refactor: move redux stuff into redux dir & api into utils
