+++ src/pages/data-set-report/DataSetReport.js
@@ -13,7 +13,7 @@ import XLSX from 'xlsx';
 
 /* Redux */
 import { connect } from 'react-redux';
-import { updateFeedbackState } from '../../actions/feedback';
+import { updateFeedbackState } from '../../redux/actions/feedback';
 
 /* i18n */
 import i18n from '../../locales';
