+++ src/redux/reducers/periodTypes.js
