+++ src/redux/reducers/standardReport.js
@@ -2,11 +2,11 @@ import { actionTypes } from '../actions/standardReport';
 import {
     ADD_NEW_REPORT_ACTION,
     CONTEXT_MENU_ACTION,
-} from '../pages/standard-report/standard.report.conf';
-import { INITIAL_PAGER } from '../utils/pagination';
-import { ACTION_MESSAGE, ERROR, LOADING, SUCCESS } from '../utils/feedbackSnackBarTypes';
-import i18n from '../locales';
-import { i18nKeys } from '../i18n';
+} from '../../pages/standard-report/standard.report.conf';
+import { INITIAL_PAGER } from '../../utils/pagination';
+import { ACTION_MESSAGE, ERROR, LOADING, SUCCESS } from '../../utils/feedbackSnackBarTypes';
+import i18n from '../../locales';
+import { i18nKeys } from '../../i18n';
 
 const defaultState = {
     pager: INITIAL_PAGER,
