+++ src/redux/selectors/periodTypes.js
